@@ -3,11 +3,16 @@ version: 0.0.1+1
 publish_to: none
 description: Your space to connect
 environment: 
-  sdk: '>=3.1.0 <4.0.0'
+  sdk: '>=3.6.0 <4.0.0'
+resolution: workspace
 
 dependencies: 
   build: any
   build_config: ^1.1.1
+  commet_calendar_widget:
+    path: ../widgets/calendar
+  matrix_widget_api:
+    path: ../widgets/matrix_widget_api
   crypto: ^3.0.2
   desktop_drop: ^0.4.1
   desktop_notifications: ^0.6.3
@@ -109,7 +114,9 @@ dependencies:
   webrtc_interface: any
   uuid: any
   flutter_background: ^1.3.0+1
+  icalendar_parser: ^2.1.0
   image_picker: ^1.2.1
+  timezone: ^0.10.1
 dependency_overrides:
   analyzer: ^7.3.0
 
