@@ -4,8 +4,9 @@ description: A new Flutter project.
 publish_to: 'none'
 version: 1.0.0+1
 
-environment:
-  sdk: '>=3.1.0 <4.0.0'
+environment: 
+  sdk: '>=3.6.0 <4.0.0'
+resolution: workspace
 
 dependencies:
   flutter:
@@ -25,7 +26,6 @@ dev_dependencies:
   build_runner:
   widgetbook_generator: ^3.1.0
   widgetbook:
-  flutter_lints: ^2.0.0
 
 flutter:
   uses-material-design: true
