@@ -1,18 +1,25 @@
 import 'dart:async';
 
+import 'package:commet/client/components/calendar_room/calendar_room_component.dart';
 import 'package:commet/client/components/voip_room/voip_room_component.dart';
 import 'package:commet/client/room.dart';
 import 'package:commet/main.dart';
 import 'package:commet/ui/atoms/adaptive_context_menu.dart';
 import 'package:commet/ui/atoms/dot_indicator.dart';
 import 'package:commet/ui/atoms/notification_badge.dart';
+import 'package:commet/ui/atoms/tiny_pill.dart';
+import 'package:commet_calendar_widget/calendar.dart';
 import 'package:flutter/material.dart';
 import 'package:tiamat/atoms/context_menu.dart';
 import 'package:tiamat/tiamat.dart' as tiamat;
 
 class RoomTextButton extends StatefulWidget {
-  const RoomTextButton(this.room,
-      {this.highlight = false, this.onTap, super.key});
+  const RoomTextButton(
+    this.room, {
+    this.highlight = false,
+    this.onTap,
+    super.key,
+  });
   final bool highlight;
   final Room room;
   final Function(Room room)? onTap;
@@ -24,19 +31,29 @@ class RoomTextButton extends StatefulWidget {
 class _RoomTextButtonState extends State<RoomTextButton> {
   late List<StreamSubscription> subs;
   VoipRoomComponent? voipRoom;
+  CalendarRoom? calendarRoom;
   List<String>? voipRoomParticipants;
+  List<MatrixCalendarEventState>? calendarEvents;
 
   @override
   void initState() {
     voipRoom = widget.room.getComponent<VoipRoomComponent>();
+    calendarRoom = widget.room.getComponent<CalendarRoom>();
+
     subs = [
       widget.room.onUpdate.listen(onRoomUpdate),
-      if (voipRoom?.isVoipRoom == true)
+      if (voipRoom != null)
         voipRoom!.onParticipantsChanged.listen(onVoipParticipantsChanged),
+      if (calendarRoom != null)
+        calendarRoom!.onEventsChanged.listen(onCalendarEventsChanged),
     ];
 
-    if (voipRoom?.isVoipRoom == true) {
+    if (voipRoom != null) {
       voipRoomParticipants = voipRoom?.getCurrentParticipants();
+    }
+
+    if (calendarRoom?.calendar != null) {
+      onCalendarEventsChanged(());
     }
 
     if (voipRoomParticipants?.isNotEmpty == true) {
@@ -58,6 +75,15 @@ class _RoomTextButtonState extends State<RoomTextButton> {
       sub.cancel();
     }
     super.dispose();
+  }
+
+  void onCalendarEventsChanged(void event) {
+    setState(() {
+      calendarEvents = calendarRoom!
+          .getEventsOnDay(DateTime.now())
+          .where((i) => i.isUnavailability == false)
+          .toList();
+    });
   }
 
   void onRoomUpdate(void event) {
@@ -90,6 +116,10 @@ class _RoomTextButtonState extends State<RoomTextButton> {
       customBuilder = buildCallParticipants;
     }
 
+    if (calendarEvents?.isNotEmpty == true) {
+      customBuilder = buildEvents;
+    }
+
     Widget result = SizedBox(
       height: customBuilder == null ? height : null,
       child: tiamat.TextButton(
@@ -119,21 +149,23 @@ class _RoomTextButtonState extends State<RoomTextButton> {
             ? NotificationBadge(widget.room.displayHighlightedNotificationCount)
             : widget.room.displayNotificationCount > 0
                 ? const Padding(
-                    padding: EdgeInsets.all(2.0),
-                    child: DotIndicator(),
-                  )
+                    padding: EdgeInsets.all(2.0), child: DotIndicator())
                 : null,
       ),
     );
 
-    if (voipRoom?.isVoipRoom == true) {
-      result = AdaptiveContextMenu(items: [
-        if (preferences.developerMode)
-          ContextMenuItem(
+    if (voipRoom != null) {
+      result = AdaptiveContextMenu(
+        items: [
+          if (preferences.developerMode)
+            ContextMenuItem(
               text: "Clear Membership Status",
               icon: Icons.call_end,
-              onPressed: () => voipRoom?.clearAllCallMembershipStatus())
-      ], child: result);
+              onPressed: () => voipRoom?.clearAllCallMembershipStatus(),
+            ),
+        ],
+        child: result,
+      );
     }
 
     return result;
@@ -152,7 +184,36 @@ class _RoomTextButtonState extends State<RoomTextButton> {
                 buildCallMember(participant),
             ],
           ),
-        )
+        ),
+      ],
+    );
+  }
+
+  Widget buildEvents(Widget child, BuildContext context) {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        SizedBox(height: height, child: child),
+        Padding(
+          padding: const EdgeInsets.fromLTRB(28, 0, 0, 4),
+          child: DecoratedBox(
+            decoration: BoxDecoration(
+              color: Theme.of(context).colorScheme.surfaceDim.withAlpha(180),
+              borderRadius: BorderRadius.circular(8),
+            ),
+            child: Padding(
+              padding: const EdgeInsets.all(8.0),
+              child: Wrap(
+                spacing: 4,
+                runSpacing: 4,
+                children: [
+                  tiamat.Text.labelLow("Today: "),
+                  for (var event in calendarEvents!) buildEvent(event),
+                ],
+              ),
+            ),
+          ),
+        ),
       ],
     );
   }
@@ -166,14 +227,32 @@ class _RoomTextButtonState extends State<RoomTextButton> {
     }
 
     return SizedBox(
-        height: height,
-        child: tiamat.TextButton(
-          member.displayName,
-          textColor: color,
-          avatar: member.avatar,
-          avatarPlaceholderColor: member.defaultColor,
-          avatarPlaceholderText: member.displayName,
-        ));
+      height: height,
+      child: tiamat.TextButton(
+        member.displayName,
+        textColor: color,
+        avatar: member.avatar,
+        avatarPlaceholderColor: member.defaultColor,
+        avatarPlaceholderText: member.displayName,
+      ),
+    );
+  }
+
+  Widget buildEvent(MatrixCalendarEventState event) {
+    var color =
+        calendarRoom!.calendar!.config.getColorFromUser(event.senderId!);
+
+    return TinyPill(
+      event.data.title,
+      background: calendarRoom!.calendar!.config.processEventColor(
+        color,
+        context,
+      ),
+      foreground: calendarRoom!.calendar!.config.processEventTextColor(
+        color,
+        context,
+      ),
+    );
   }
 
   void onVoipParticipantsChanged(void event) {
