@@ -14,11 +14,11 @@ class TinyPill extends StatelessWidget {
       padding: const EdgeInsets.fromLTRB(0, 0, 0, 0),
       child: Container(
         decoration: BoxDecoration(
-            borderRadius: BorderRadius.circular(5),
-            color:
-                background ?? Theme.of(context).colorScheme.primaryContainer),
+          borderRadius: BorderRadius.circular(5),
+          color: background ?? Theme.of(context).colorScheme.primaryContainer,
+        ),
         child: Padding(
-          padding: const EdgeInsets.all(2.0),
+          padding: const EdgeInsets.fromLTRB(4, 2, 4, 2),
           child: tiamat.Text.tiny(
             text,
             color:
