@@ -1,4 +1,5 @@
 import 'package:commet/config/experiments.dart';
+import 'package:commet/ui/pages/settings/categories/app/general_settings_page.dart';
 import 'package:flutter/material.dart';
 
 import 'package:tiamat/tiamat.dart' as tiamat;
@@ -21,91 +22,65 @@ class _ExperimentsSettingsPageState extends State<ExperimentsSettingsPage> {
         const Padding(
           padding: EdgeInsets.all(8.0),
           child: tiamat.Text.label(
-              "These features are still under development, and may contain bugs or security issues. Enable at your own risk"),
+            "These features are still under development, and may contain bugs or security issues. Enable at your own risk",
+          ),
         ),
         Panel(
           header: "Experiments",
           mode: TileType.surfaceContainerLow,
-          child: Column(children: [
-            Row(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              children: [
-                const Flexible(
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      tiamat.Text.labelEmphasised("1:1 Voice and Video Calls"),
-                      tiamat.Text.labelLow(
-                          "Enables peer to peer voice and video calls, using WebRTC")
-                    ],
-                  ),
-                ),
-                tiamat.Switch(
-                  state: Experiments.voip,
-                  onChanged: (value) async {
-                    await Experiments.setVoip(value);
-                    setState(() {});
-                  },
-                )
-              ],
-            ),
-            SizedBox(
-              height: 20,
-            ),
-            Row(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              children: [
-                const Flexible(
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      tiamat.Text.labelEmphasised(
-                          "Voice/Video Rooms (Element Call)"),
-                      tiamat.Text.labelLow(
-                          "Enables group video and audio calls using LiveKit")
-                    ],
-                  ),
-                ),
-                tiamat.Switch(
-                  state: Experiments.elementCall,
-                  onChanged: (value) async {
-                    await Experiments.setElementCall(value);
-                    setState(() {});
-                  },
-                )
-              ],
-            ),
-            SizedBox(
-              height: 20,
-            ),
-            Row(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              children: [
-                const Flexible(
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      tiamat.Text.labelEmphasised("Photo Album Rooms"),
-                      tiamat.Text.labelLow(
-                          "Share photos and videos, with dedicated album viewer")
-                    ],
-                  ),
-                ),
-                tiamat.Switch(
-                  state: Experiments.photoAlbumRooms,
-                  onChanged: (value) async {
-                    await Experiments.setPhotoAlbumRooms(value);
-                    setState(() {});
-                  },
-                )
-              ],
-            )
-          ]),
+          child: Column(
+            children: [
+              GeneralSettingsPageState.settingToggle(
+                Experiments.voip,
+                title: "1:1 Voice and Video Calls",
+                description:
+                    "Enables peer to peer voice and video calls, using WebRTC",
+                onChanged: (value) async {
+                  setState(() {
+                    Experiments.setVoip(value);
+                  });
+                },
+              ),
+              GeneralSettingsPageState.settingToggle(
+                Experiments.elementCall,
+                title: "Voice/Video Rooms (Element Call)",
+                description:
+                    "Enables group video and audio calls using LiveKit",
+                onChanged: (value) async {
+                  setState(() {
+                    Experiments.setElementCall(value);
+                  });
+                },
+              ),
+              GeneralSettingsPageState.settingToggle(
+                Experiments.photoAlbumRooms,
+                title: "Photo Album Rooms",
+                description:
+                    "Share photos and videos, with dedicated album viewer",
+                onChanged: (value) async {
+                  setState(() {
+                    Experiments.setPhotoAlbumRooms(value);
+                  });
+                },
+              ),
+              GeneralSettingsPageState.settingToggle(
+                Experiments.calendarRooms,
+                title: "Calendar Rooms",
+                description: "Create a shared calendar using matrix rooms",
+                onChanged: (value) async {
+                  setState(() {
+                    Experiments.setCalendarRoom(value);
+                  });
+                },
+              ),
+            ],
+          ),
         ),
         const Padding(
           padding: EdgeInsets.all(8.0),
           child: tiamat.Text.error(
-              "You must restart the app for changes to take effect"),
+            "You must restart the app for changes to take effect",
+          ),
         ),
       ],
     );
