@@ -12,30 +12,22 @@ const String loremIpsum =
 
 @UseCase(name: 'Label', type: Text)
 Widget wbtextLabelUseCase(BuildContext context) {
-  return const material.Center(
-    child: Text.label(exampleText),
-  );
+  return const material.Center(child: Text.label(exampleText));
 }
 
 @UseCase(name: 'Tiny', type: Text)
 Widget wbtextTinyUseCase(BuildContext context) {
-  return const material.Center(
-    child: Text.tiny(exampleText),
-  );
+  return const material.Center(child: Text.tiny(exampleText));
 }
 
 @UseCase(name: 'Body', type: Text)
 Widget wbtextBodyUseCase(BuildContext context) {
-  return const material.Center(
-    child: Text.body(exampleText),
-  );
+  return const material.Center(child: Text.body(exampleText));
 }
 
 @UseCase(name: 'Error', type: Text)
 Widget wbtextErrorUseCase(BuildContext context) {
-  return const material.Center(
-    child: Text.error(exampleText),
-  );
+  return const material.Center(child: Text.error(exampleText));
 }
 
 @UseCase(name: 'Title', type: Text)
@@ -53,10 +45,7 @@ Widget wbtextNameUseCase(BuildContext context) {
   return const material.Center(
     child: material.Padding(
       padding: EdgeInsets.all(8.0),
-      child: Text.name(
-        exampleText,
-        color: Colors.amber,
-      ),
+      child: Text.name(exampleText, color: Colors.amber),
     ),
   );
 }
@@ -74,23 +63,30 @@ Widget wbtextAllUseCase(BuildContext context) {
           Text.largeTitle(exampleText),
           Seperator(),
           Align(
-              alignment: Alignment.centerRight,
-              child: Text.labelEmphasised(exampleText)),
+            alignment: Alignment.centerRight,
+            child: Text.labelEmphasised(exampleText),
+          ),
           Align(
-              alignment: Alignment.centerRight, child: Text.label(exampleText)),
+            alignment: Alignment.centerRight,
+            child: Text.label(exampleText),
+          ),
           material.Padding(
             padding: EdgeInsets.fromLTRB(0, 8, 0, 8),
             child: Text.body(
-                "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum."),
+              "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.",
+            ),
           ),
           material.Padding(
             padding: EdgeInsets.fromLTRB(0, 8, 0, 8),
             child: Text.tiny(
-                "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua"),
+              "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua",
+            ),
           ),
           Seperator(),
           Align(
-              alignment: Alignment.centerRight, child: Text.error(exampleText)),
+            alignment: Alignment.centerRight,
+            child: Text.error(exampleText),
+          ),
         ],
       ),
     ),
@@ -105,18 +101,20 @@ enum TextType {
   tiny,
   body,
   largeTitle,
-  name
+  name,
 }
 
 class Text extends StatelessWidget {
-  const Text(this.text,
-      {super.key,
-      this.type = TextType.label,
-      this.overflow,
-      this.color,
-      this.softwrap,
-      this.autoAdjustBrightness,
-      this.maxLines});
+  const Text(
+    this.text, {
+    super.key,
+    this.type = TextType.label,
+    this.overflow,
+    this.color,
+    this.softwrap,
+    this.autoAdjustBrightness,
+    this.maxLines,
+  });
   final String text;
   final TextType type;
   final Color? color;
@@ -131,10 +129,10 @@ class Text extends StatelessWidget {
     this.overflow,
     this.maxLines,
     this.softwrap,
-  })  : type = TextType.label,
-        color = null,
-        autoAdjustBrightness = false,
-        super(key: key);
+  }) : type = TextType.label,
+       color = null,
+       autoAdjustBrightness = false,
+       super(key: key);
 
   const Text.labelEmphasised(
     this.text, {
@@ -143,9 +141,9 @@ class Text extends StatelessWidget {
     this.color,
     this.maxLines,
     this.softwrap,
-  })  : type = TextType.labelEmphasised,
-        autoAdjustBrightness = false,
-        super(key: key);
+  }) : type = TextType.labelEmphasised,
+       autoAdjustBrightness = false,
+       super(key: key);
 
   const Text.error(
     this.text, {
@@ -153,10 +151,10 @@ class Text extends StatelessWidget {
     this.overflow,
     this.maxLines,
     this.softwrap,
-  })  : type = TextType.error,
-        color = null,
-        autoAdjustBrightness = false,
-        super(key: key);
+  }) : type = TextType.error,
+       color = null,
+       autoAdjustBrightness = false,
+       super(key: key);
 
   const Text.tiny(
     this.text, {
@@ -165,9 +163,9 @@ class Text extends StatelessWidget {
     this.maxLines,
     this.color,
     this.softwrap,
-  })  : type = TextType.tiny,
-        autoAdjustBrightness = false,
-        super(key: key);
+  }) : type = TextType.tiny,
+       autoAdjustBrightness = false,
+       super(key: key);
 
   const Text.body(
     this.text, {
@@ -175,10 +173,10 @@ class Text extends StatelessWidget {
     this.overflow,
     this.maxLines,
     this.softwrap,
-  })  : type = TextType.body,
-        color = null,
-        autoAdjustBrightness = false,
-        super(key: key);
+  }) : type = TextType.body,
+       color = null,
+       autoAdjustBrightness = false,
+       super(key: key);
 
   const Text.largeTitle(
     this.text, {
@@ -186,10 +184,10 @@ class Text extends StatelessWidget {
     this.overflow,
     this.maxLines,
     this.softwrap,
-  })  : type = TextType.largeTitle,
-        color = null,
-        autoAdjustBrightness = false,
-        super(key: key);
+  }) : type = TextType.largeTitle,
+       color = null,
+       autoAdjustBrightness = false,
+       super(key: key);
 
   const Text.name(
     this.text, {
@@ -198,9 +196,9 @@ class Text extends StatelessWidget {
     this.overflow,
     this.maxLines,
     this.softwrap,
-  })  : type = TextType.name,
-        autoAdjustBrightness = true,
-        super(key: key);
+  }) : type = TextType.name,
+       autoAdjustBrightness = true,
+       super(key: key);
 
   const Text.labelLow(
     this.text, {
@@ -209,22 +207,30 @@ class Text extends StatelessWidget {
     this.overflow,
     this.maxLines,
     this.softwrap,
-  })  : type = TextType.labelLow,
-        autoAdjustBrightness = false,
-        super(key: key);
+  }) : type = TextType.labelLow,
+       autoAdjustBrightness = false,
+       super(key: key);
 
-  static Color adjustColor(BuildContext context, Color color) {
+  static Color adjustColor(
+    BuildContext context,
+    Color color, {
+    double saturationMultiplier = 1,
+  }) {
     var hsl = HSLColor.fromColor(color!);
     double lightness = hsl.lightness;
-    double saturation = hsl.saturation;
+    double saturation = hsl.saturation * saturationMultiplier;
     if (Theme.of(context).brightness == Brightness.dark) {
       lightness = clampDouble(hsl.lightness, 0.75, 1);
     } else {
       lightness = clampDouble(hsl.lightness, 0, 0.7);
     }
 
-    return HSLColor.fromAHSL(hsl.alpha, hsl.hue, saturation, lightness)
-        .toColor();
+    return HSLColor.fromAHSL(
+      hsl.alpha,
+      hsl.hue,
+      saturation,
+      lightness,
+    ).toColor();
   }
 
   @override
@@ -237,47 +243,54 @@ class Text extends StatelessWidget {
 
     switch (type) {
       case TextType.label:
-        style = material.Theme.of(context)
-            .textTheme
-            .labelLarge!
-            .copyWith(fontWeight: FontWeight.w300, color: newColor);
+        style = material.Theme.of(context).textTheme.labelLarge!.copyWith(
+          fontWeight: FontWeight.w300,
+          color: newColor,
+        );
         break;
       case TextType.labelEmphasised:
-        style = material.Theme.of(context)
-            .textTheme
-            .labelLarge!
-            .copyWith(fontWeight: FontWeight.w400, color: newColor);
+        style = material.Theme.of(context).textTheme.labelLarge!.copyWith(
+          fontWeight: FontWeight.w400,
+          color: newColor,
+        );
         break;
       case TextType.error:
         style = material.Theme.of(context).textTheme.bodyMedium!.copyWith(
-            fontWeight: FontWeight.w500,
-            color: newColor ?? Theme.of(context).colorScheme.error);
+          fontWeight: FontWeight.w500,
+          color: newColor ?? Theme.of(context).colorScheme.error,
+        );
         break;
       case TextType.tiny:
         style = material.Theme.of(context).textTheme.bodyMedium!.copyWith(
-            fontWeight: FontWeight.w300, fontSize: 10, color: newColor);
+          fontWeight: FontWeight.w300,
+          fontSize: 10,
+          color: newColor,
+        );
         break;
       case TextType.body:
-        style = material.Theme.of(context)
-            .textTheme
-            .bodyMedium!
-            .copyWith(fontWeight: FontWeight.w300, color: newColor);
+        style = material.Theme.of(context).textTheme.bodyMedium!.copyWith(
+          fontWeight: FontWeight.w300,
+          color: newColor,
+        );
         break;
       case TextType.largeTitle:
-        style = material.Theme.of(context)
-            .textTheme
-            .titleLarge!
-            .copyWith(color: newColor);
+        style = material.Theme.of(
+          context,
+        ).textTheme.titleLarge!.copyWith(color: newColor);
         break;
       case TextType.name:
         style = material.Theme.of(context).textTheme.bodyMedium!.copyWith(
-            color: newColor, fontWeight: FontWeight.w400, fontSize: 15);
+          color: newColor,
+          fontWeight: FontWeight.w400,
+          fontSize: 15,
+        );
         break;
       case TextType.labelLow:
         style = material.Theme.of(context).textTheme.bodyMedium!.copyWith(
-            color: newColor ?? Theme.of(context).colorScheme.secondary,
-            fontWeight: FontWeight.w400,
-            fontSize: 12);
+          color: newColor ?? Theme.of(context).colorScheme.secondary,
+          fontWeight: FontWeight.w400,
+          fontSize: 12,
+        );
         break;
     }
 
