@@ -9,8 +9,11 @@ import 'package:tiamat/config/style/theme_dark.dart';
 class ThemeYou {
   static Future<ThemeData> theme(Brightness brightness) async {
     ColorScheme? scheme;
-    scheme = (await DynamicColorPlugin.getColorScheme())?[
-        brightness == Brightness.light ? 0 : 1];
+    scheme =
+        (await DynamicColorPlugin.getColorScheme())?[brightness ==
+                Brightness.light
+            ? 0
+            : 1];
 
     if (scheme == null) {
       var color = await DynamicColorPlugin.getAccentColor();
@@ -28,10 +31,26 @@ class ThemeYou {
       }
     }
 
-    return ThemeBase.theme(scheme).copyWith(extensions: [
-      const ThemeSettings(),
-      ExtraColors.fromScheme(scheme),
-      FoundationSettings(color: scheme.surfaceDim)
-    ]);
+    return ThemeBase.theme(scheme).copyWith(
+      extensions: [
+        const ThemeSettings(),
+        ExtraColors.fromScheme(scheme),
+        FoundationSettings(color: scheme.surfaceDim),
+      ],
+    );
+  }
+
+  static ThemeData withSeedColor(Brightness brightness, Color color) {
+    ColorScheme? scheme;
+
+    scheme = ColorScheme.fromSeed(seedColor: color, brightness: brightness);
+
+    return ThemeBase.theme(scheme).copyWith(
+      extensions: [
+        const ThemeSettings(),
+        ExtraColors.fromScheme(scheme),
+        FoundationSettings(color: scheme.surfaceDim),
+      ],
+    );
   }
 }
